Created ArgChecker.py
